Web Service Tester, car and transaction Service implemented. Adding Documentation and comments.

+++ src/classes/Client.java
@@ -31,16 +31,13 @@ public class Client {
         this.age = age;
     }
 
-    //private String name;
-
-    public long getId_client() {
+    public int getId_client() {
         return id_client;
     }
 
     public void setId_client(int id_client) {
         this.id_client = id_client;
     }
-
 
     public Boolean getMembership() {
         return membership;
